Components: Extract Reusable Post Sticky Component
